requirements.txt
aiohttp==3.9.4
aiosignal==1.3.1
alabaster==0.7.16
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
Babel==2.14.0
certifi==2024.2.2
cftime==1.6.3
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
cmake==3.29.2
dask==2024.4.1
dgl==2.1.0+cu117
docutils==0.21.1
exceptiongroup==1.2.0
filelock==3.13.4
frozenlist==1.4.1
fsspec==2024.3.1
idna==3.7
imagesize==1.4.1
importlib_metadata==7.1.0
iniconfig==2.0.0
Jinja2==3.1.3
joblib==1.4.0
lit==18.1.3
locket==1.0.0
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.0.5
netCDF4==1.6.5
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
packaging==24.0
pandas==2.2.2
partd==1.4.1
pillow==10.3.0
pluggy==1.4.0
portalocker==2.8.2
psutil==5.9.8
pydantic==2.7.0
pydantic_core==2.18.1
Pygments==2.17.2
pyparsing==3.1.2
pytest==8.1.1
python-dateutil==2.9.0.post0
python-sat==1.8.dev12
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
snowballstemmer==2.2.0
Sphinx==7.3.7
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
sympy==1.12
threadpoolctl==3.4.0
tomli==2.0.1
toolz==0.12.1
torch==2.0.1
torch_geometric==2.5.2
torchaudio==2.0.2
torchdata==0.7.1
torchvision==0.15.2
tqdm==4.66.2
triton==2.0.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
xarray==2024.3.0
yarl==1.9.4
zipp==3.18.1
